Move podman logic to src/podman
