+++ cmd/podman/status.go
@@ -1,10 +1,8 @@
 package podman
 
 import (
-	"fmt"
 	"github.com/spf13/cobra"
-	"infra-lab-cli/config"
-	"infra-lab-cli/utils"
+	podmansrc "infra-lab-cli/src/podman"
 )
 
 var StatusCmd = &cobra.Command{
@@ -13,18 +11,6 @@ var StatusCmd = &cobra.Command{
 	RunE:  runStatus,
 }
 
-func runStatus(cmd *cobra.Command, args []string) (err error) {
-	if !config.IsBinaryInPath(binaryName) {
-		fmt.Print(config.BinaryNotFoundError(binaryName))
-		return nil
-	}
-
-	err = InspectCmd.RunE(InspectCmd, args)
-	if err != nil {
-		return err
-	}
-	fmt.Printf("%s\t %s\t %d cpu\t %.1f GiB\t %d GiB\n",
-		inspectedMachines[0].Name, inspectedMachines[0].State,
-		inspectedMachines[0].Resources.CPUs, utils.ConvertMiBToGiB(inspectedMachines[0].Resources.Memory), inspectedMachines[0].Resources.DiskSize)
-	return nil
+func runStatus(cmd *cobra.Command, args []string) error {
+	return podmansrc.GetMachineStatus(binaryName, machineName)
 }
