+++ src/podman/types.go
@@ -1,0 +1,58 @@
+package podman
+
+// ListedMachine represents a machine from the list command
+type ListedMachine struct {
+	Name               string `json:"Name"`
+	Default            bool   `json:"Default"`
+	Created            string `json:"Created"`
+	Running            bool   `json:"Running"`
+	Starting           bool   `json:"Starting"`
+	LastUp             string `json:"LastUp"`
+	Stream             string `json:"Stream"`
+	VMType             string `json:"VMType"`
+	CPUs               int    `json:"CPUs"`
+	Memory             string `json:"Memory"`
+	DiskSize           string `json:"DiskSize"`
+	Port               int    `json:"Port"`
+	RemoteUsername     string `json:"RemoteUsername"`
+	IdentityPath       string `json:"IdentityPath"`
+	UserModeNetworking bool   `json:"UserModeNetworking"`
+}
+
+// InspectedMachine represents a machine from the inspect command
+type InspectedMachine struct {
+	ConfigDir struct {
+		Path string `json:"Path"`
+	} `json:"ConfigDir"`
+	ConnectionInfo struct {
+		PodmanSocket struct {
+			Path string `json:"Path"`
+		} `json:"PodmanSocket"`
+		PodmanPipe interface{} `json:"PodmanPipe"`
+	} `json:"ConnectionInfo"`
+	Created   string `json:"Created"`
+	LastUp    string `json:"LastUp"`
+	Name      string `json:"Name"`
+	Resources struct {
+		CPUs     int           `json:"CPUs"`
+		DiskSize int           `json:"DiskSize"`
+		Memory   int           `json:"Memory"`
+		USBs     []interface{} `json:"USBs"`
+	} `json:"Resources"`
+	SSHConfig struct {
+		IdentityPath   string `json:"IdentityPath"`
+		Port           int    `json:"Port"`
+		RemoteUsername string `json:"RemoteUsername"`
+	} `json:"SSHConfig"`
+	State              string `json:"State"`
+	UserModeNetworking bool   `json:"UserModeNetworking"`
+	Rootful            bool   `json:"Rootful"`
+	Rosetta            bool   `json:"Rosetta"`
+}
+
+// MachineField represents a field for displaying machine information
+type MachineField struct {
+	Name   string
+	Header string
+	Width  int
+}
